Get actions from the Trello API
The actions hold data about the movement of cards. When a card is moved
to another list, the board actions are updated.

Use the action to get the date a card was moved to a list, and to
calculate the number of days it has been since the card was moved.

The list name is currently hard-coded to “List One” for the tests. This
will be updated to Factcheck in forthcoming commits