Renames TrelloAPI into Trello
This is a prepatory commit. The intention is to move all the logic that
need to know about the data returned by Trello into Trello.js. ie. search 
for a Zendesk attachment, the days in the list…

This will also facilitate reusing in other reports, although this is not
what is driving this change.