Move findZendeskId method to Trello
Testing is now much easier. The unit test only involves the card, and
the report test can rely on a Spy on the Trello object because it is
already tested.

I would need to get rid of the fixture, but as it is used in other tests
I will wait until all the code has been moved there.