Move boardUrl out of Trello class
This is a preparatory commit to make accessible to the report to a
pop-up menu in the Chrome extension.

The Chrome extension pop-up run the Factcheck report in its own html.
For this reason `window.location.pathname` is not returning the board
url, but the popup url. 

By extracting this information out of the Trello object we prepare
next popup so it can pass the URL as a parameter.